lib/transpiled: add arm64 copies for darwin and windows
Given that the only difference between architectures has proven to be
use of inline assembly for SIMD acceleration, we can simply copy the
transpiled output for arm64 variants.

Ideally, we would setup a full cross compile for these artifacts, though
that is proving to be a total pain. If this causes issues we can revert.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>